feat(Dashboard): add sidebar with navigation to dashboard layout
