+++ src/app/login/page.tsx
@@ -19,7 +19,7 @@ const LoginPage = () => {
 
   useEffect(() => {
     if (!loading && user) {
-      router.replace("/");
+      router.replace("/dashboard");
     }
   }, [user, loading, router]);
